fix: pr 피드백 반영
