+++ src/mocks/browsers.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw';
-import { dataHandlers } from './dataHandlers';
+import { dataHandlers } from './boardHandlers';
 import { handlers } from './handlers';
 import { workspaceHandlers } from './workspaceHandlers';
